feat(codegen): bundle whitelist entries within codegen
