+++ examples/vite/package.json
@@ -9,7 +9,6 @@
     "preview": "vite preview"
   },
   "devDependencies": {
-    "@polkadot-api/cli": "workspace:*",
     "@types/node": "^20.4.1",
     "ts-node": "^10.9.1",
     "typescript": "^5.3.2",
