+++ packages/codegen/src/index.ts
@@ -3,4 +3,3 @@ export * from "./get-new-types"
 export * from "./generate-descriptors"
 export * from "./generate-multiple-descriptors"
 export { knownTypesRepository } from "./known-types"
-export type { Whitelist } from "./whitelist"
